requirements.txt
Django==2.1.2
django-haystack==2.8.1
elasticsearch==2.4.1
pytz==2018.6
selenium==3.14.1
urllib3==1.24
